requirement.txt
transformers==4.37.2
peft==0.3.0
accelerate==0.25.0
torch>=2.0.0
torchaudio>=2.0.0
python-multipart==0.0.9
pydantic==2.6.3
numpy==1.24.4
scikit-learn==1.3.2
librosa==0.10.1
matplotlib==3.8.3
soundfile==0.12.1
streamlit
Pillow
opencv-python-headless==4.5.1.48
plotly
requests
tensorflow-cpu
